backend/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
redis==5.0.8
duckduckgo-search==6.3.4
google-generativeai==0.7.2
requests==2.32.3
